added class unit tests - all working
