+++ pyrealm/constants/two_leaf_canopy.py
@@ -1,8 +1,6 @@
 """To do."""
 
-from dataclasses import dataclass, field
-
-import numpy as np
+from dataclasses import dataclass
 
 from pyrealm.constants import ConstantsClass
 
@@ -24,5 +22,5 @@ class TwoLeafConst(ConstantsClass):
     k_kd_prime: float = 0.719  # needs citation
     """diffuse and scattered diffuse PAR extinction coefficient, dimensionless"""
 
-    k_sol_obs_angle: float = field(default_factory=np.deg2rad(1))  # needs a citation
+    k_sol_obs_angle: float = 0.01745329  # 1 degree in rads, needs a citation
     """ solar obscurity angle, radians"""
